code

+++ file/src/main/java/com/dtstack/logstash/inputs/ReadFactory.java
@@ -13,7 +13,7 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class ReadFactory {
 	
-	public static IReader createReader(java.io.File file, String encoding, ConcurrentHashMap<String, Integer> fileCurrPos
+	public static IReader createReader(java.io.File file, String encoding, ConcurrentHashMap<String, Long> fileCurrPos
 			, String startPos) throws IOException{
 		
 		IReader reader = null;
@@ -25,7 +25,7 @@ public class ReadFactory {
 		}else if(fileName.toLowerCase().endsWith(".tar.gz") || fileName.toLowerCase().endsWith(".tar")){
 			reader = ReadTarFile.createInstance(fileName, encoding, fileCurrPos);
 		}else{
-			Integer filePos = fileCurrPos.get(fileName);
+			Long filePos = fileCurrPos.get(fileName);
 			
 			if(filePos == null){//未读取过的根据配置的读取点开始
 				reader = new ReadLineUtil(file, encoding, startPos);
